remove function from navigation and keep login in the component
