+++ app/screens/home/search/team_picker_icon/team_picker_icon.tsx
@@ -7,10 +7,11 @@ import {View, useWindowDimensions} from 'react-native';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 
 import CompassIcon from '@components/compass_icon';
+import {getTeamsSnapHeight} from '@components/team_sidebar/add_team';
 import TeamIcon from '@components/team_sidebar/team_list/team_item/team_icon';
 import TouchableWithFeedback from '@components/touchable_with_feedback';
 import {useTheme} from '@context/theme';
-import {bottomSheetWithTeamList} from '@screens/navigation';
+import {bottomSheet} from '@screens/navigation';
 import {preventDoubleTap} from '@utils/tap';
 import {changeOpacity, makeStyleSheetFromTheme} from '@utils/theme';
 
@@ -71,13 +72,13 @@ const TeamPickerIcon = ({size = 24, divider = false, setTeamId, teams, teamId}: 
             );
         };
 
-        bottomSheetWithTeamList({
-            dimensions,
-            insets,
+        const height = getTeamsSnapHeight({dimensions, teams, insets});
+        bottomSheet({
+            closeButtonId: 'close-team_list',
             renderContent,
+            snapPoints: [height, 10],
             theme,
             title,
-            teams,
         });
     }), [theme, setTeamId, teamId, teams]);
